refactor: Lotto 구매에 알맞은 단위 Money는 모르도록 수정
