+++ src/main/java/lotto/model/Lotto.java
@@ -7,6 +7,7 @@ import lotto.model.number.LottoNumbers;
  * 자동 로또 한 장을 의미하는 Class
  */
 public class Lotto {
+    private static final String ERROR_UNIT = "[ERROR] 구매 금액은 1000원 단위로 입력하세요";
     private static final int PRICE = 1000;
 
     private LottoNumbers numbers;
@@ -16,7 +17,14 @@ public class Lotto {
     }
 
     public static int countAvailableTickets(Money money) {
+        checkUnit(money);
         return money.countAvailable(PRICE);
+    }
+
+    private static void checkUnit(Money money) {
+        if (!money.isUnit(PRICE)) {
+            throw new IllegalArgumentException(ERROR_UNIT);
+        }
     }
 
     public boolean contains(LottoNumber number) {
