+++ src/main/java/lotto/model/Money.java
@@ -2,16 +2,12 @@ package lotto.model;
 
 public class Money {
     private static final String ERROR_NEGATIVE = "[ERROR] 구매 금액은 0원보다 커야 합니다";
-    private static final String ERROR_UNIT = "[ERROR] 구매 금액은 1000원 단위로 입력하세요";
     private static final String ERROR_TYPE = "[ERROR] 구매 금액은 숫자로만 입력하세요";
-    private static final int UNIT = 1000;
 
     private final int amount;
 
     private Money(int amount) {
         checkNegative(amount);
-        checkUnit(amount);
-
         this.amount = amount;
     }
 
@@ -21,18 +17,16 @@ public class Money {
         }
     }
 
-    private void checkUnit(int amount) {
-        if (amount % UNIT != 0) {
-            throw new IllegalArgumentException(ERROR_UNIT);
-        }
-    }
-
     public static Money from(String input) {
         try {
             return new Money(Integer.parseInt(input));
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException(ERROR_TYPE);
         }
+    }
+
+    public boolean isUnit(int unit) {
+        return amount % unit == 0;
     }
 
     public double rate(int numerator) {
