+++ src/test/java/lotto/model/MoneyTest.java
@@ -24,14 +24,6 @@ public class MoneyTest {
                 .hasMessage("[ERROR] 구매 금액은 0원보다 커야 합니다");
     }
 
-    @DisplayName("구매 금액이 1000원 단위가 아니면 예외 발생")
-    @Test
-    void unit_exception() {
-        assertThatThrownBy(() -> Money.from("1500"))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 구매 금액은 1000원 단위로 입력하세요");
-    }
-
     @DisplayName("2000원치 로또가 2장인지 확인한다")
     @Test
     public void purchase_amount_2000() {
